ship_detection/requirements.txt
Pillow==9.5.0
torch==1.11.0
huggingface
transformers
xformers 
datasets[vision]
easyimages 
pybboxes 
wandb
ultralytics